starsightcrm/requirement.txt
asgiref==3.3.1
autopep8==1.5.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-crispy-forms==1.10.0
django-formtools==2.2
django-import-export==2.5.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.5
PyYAML==5.3.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
